# ignore redirects, accept more status codes for successful probes