@TAbdiukov 2024-07-08
UPDATE: Never mind, you want to follow redirects, but show this information, but show it to the user, right? (Instead of ignoring them?)