@pjperez 2024-07-09
Hey @TAbdiukov thanks a lot for this! I think the ideal scenario is one where we do not follow redirects, report the probes as successful (while showing to user the return code) and have a flag to follow redirects.