@TAbdiukov 2024-07-09
Hey @pjperez . I see,
 1. "we do not follow redirects" - done
 2. "report the probes as successful  (while showing to user the return code" - done
     - The user can probably see whether the page wanted to redirect or not based on the HTTP status code

3. "flag to follow redirects" - todo