@TAbdiukov 2024-07-09
Hey @pjperez . Thanks for your response,
 1. "we do not follow redirects" - **Done yesterday**
 2. "report the probes as successful  (while showing to user the return code" - **Done yesterday**
     - The user can probably see whether the page wanted to redirect or not based on the HTTP status code (e.g., 200 vs 302)
3. "flag to follow redirects" - **Done today**

Also when you get a chance, please check [https://github.com/pjperez/gotcping/pull/1](https://github.com/pjperez/gotcping/pull/1)

-----------------------------------------------------------------

Testing,

```

C:\Projects\httping>go run httping.go

httping 0.10.0 - A tool to measure RTT on HTTP/S requests
Help: httping -h
Usage of httping.exe:
  -count int
        Number of requests to send [0 means infinite] (default 10)
  -followredirects
        If true, follows redirects, which may result in higher RTT
  -hostheader string
        Optional: Host header
  -httpverb string
        HTTP Verb: Only GET or HEAD supported at the moment (default "GET")
  -json
        If true, produces output in json format
  -listen int
        Enable listener mode on specified port, e.g. '-r 80'
  -noproxy
        If true, ignores system proxy settings
  -timeout int
        Timeout in milliseconds (default 2000)
  -url string
        Requested URL

You haven't specified a URL to test!

exit status 1

C:\Projects\httping>go run httping.go -url http://tinyurl.com/redirect.php?num=cd77w8b -noproxy -count 4 -timeout 10000

httping 0.10.0 - A tool to measure RTT on HTTP/S requests
Help: httping -h
HTTP GET to tinyurl.com (http://tinyurl.com/redirect.php?num=cd77w8b):
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=1, httpVerb=GET, httpStatus=302, bytes=1318, RTT=335.85 ms
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=2, httpVerb=GET, httpStatus=302, bytes=1318, RTT=305.19 ms
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=3, httpVerb=GET, httpStatus=302, bytes=1318, RTT=289.74 ms
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=4, httpVerb=GET, httpStatus=302, bytes=1318, RTT=289.81 ms

Probes sent: 4
Successful responses: 4
% of requests failed: 0
Min response time: 289.7393ms
Average response time: 305.14705ms
Median response time: 297.49845ms
Max response time: 335.852ms

90% of requests were faster than: 320.51855ms
75% of requests were faster than: 305.1851ms
50% of requests were faster than: 289.8118ms
25% of requests were faster than: 289.7393ms

C:\Projects\httping>go run httping.go -url http://tinyurl.com/redirect.php?num=cd77w8b -noproxy -count 4 -timeout 10000 -followredirects

httping 0.10.0 - A tool to measure RTT on HTTP/S requests
Help: httping -h
HTTP GET to tinyurl.com (http://tinyurl.com/redirect.php?num=cd77w8b):
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=1, httpVerb=GET, httpStatus=200, bytes=1474584, RTT=2784.17 ms
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=2, httpVerb=GET, httpStatus=200, bytes=1481001, RTT=2249.17 ms
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=3, httpVerb=GET, httpStatus=200, bytes=1483976, RTT=2089.78 ms
connected to http://tinyurl.com/redirect.php?num=cd77w8b, proxy=None, seq=4, httpVerb=GET, httpStatus=200, bytes=1474238, RTT=2625.80 ms

Probes sent: 4
Successful responses: 4
% of requests failed: 0
Min response time: 1s
Average response time: 2.437232s
Median response time: 2.4374867s
Max response time: 2.7841742s

90% of requests were faster than: 2.7049886s
75% of requests were faster than: 2.625803s
50% of requests were faster than: 2.2491704s
25% of requests were faster than: 2.0897804s
```